refactor: move to service
