+++ packages/blocks/src/__internal__/utils/types.ts
@@ -28,14 +28,9 @@ export interface EditingState {
   rect: DOMRect;
 }
 
-export type DatabaseTableStateStage =
-  | 'start'
-  | 'move'
-  | 'end'
-  | 'click'
-  | 'clear';
+export type DatabaseTableStateType = 'select' | 'clear';
 export type DatabaseTableState = {
-  stage: DatabaseTableStateStage;
+  type: DatabaseTableStateType;
   databaseId?: string;
   rowIds?: number[];
 };
